requirements.txt
django~=5.0.3
djangorestframework~=3.15.1
django-rest-knox~=4.2.0